# reference/apisperu-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API REST Facturaci\u00f3n Electr\u00f3nica - APIsPERU. <br /> Si se siente m\u00e1s c\u00f3modo usando Swagger-UI, puede <a target='_blank' href='/doc/swagger/ui'>ingresar aqu\u00ed.</a>",
        "version": "1.3",
        "title": "API DE FACTURACI\u00d3N",
        "contact": {
            "name": "Apisperu",
            "email": "soporte@apisperu.com"
        },
        "x-logo": {
            "url": "/img/logoredoc.png",
            "altText": "Logo APIsPERU"
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/despatch/send": {
            "post": {
                "tags": [
                    "despatch"
                ],
                "summary": "Enviar una Gu\u00eda de Remisi\u00f3n",
                "description": "Retorna la respuesta de SUNAT y el XML firmado",
                "operationId": "sendDespatch",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Despatch"
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SummaryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de Validaci\u00f3n",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValidationResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/despatch/xml": {
            "post": {
                "tags": [
                    "despatch"
                ],
                "summary": "Generar el XML de una Gu\u00eda de Remisi\u00f3n",
                "description": "Retorna el XML",
                "operationId": "getDespatchXml",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Despatch"
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "400": {
                        "description": "Error de Validaci\u00f3n",
                        "content": {
                            "text/xml": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValidationResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/despatch/status": {
            "get": {
                "tags": [
                    "despatch"
                ],
                "summary": "Obtiene el estado del ticket",
                "description": "Retorna el CDR",
                "operationId": "getDespatchStatus",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "query",
                        "description": "Ticket del comprobante",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ruc",
                        "in": "query",
                        "description": "RUC de empresa (Multiples Empresas)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResult"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/retention/send": {
            "post": {
                "tags": [
                    "retention"
                ],
                "summary": "Enviar un Comprobante de Retencion",
                "description": "Retorna la respuesta de SUNAT y el XML firmado",
                "operationId": "sendRetention",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Retention"
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de Validaci\u00f3n",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValidationResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/retention/xml": {
            "post": {
                "tags": [
                    "retention"
                ],
                "summary": "Generar el XML de un Comprobante de Retencion",
                "description": "Retorna el XML",
                "operationId": "getRetentionXml",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Retention"
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "400": {
                        "description": "Error de Validaci\u00f3n",
                        "content": {
                            "text/xml": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValidationResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/retention/pdf": {
            "post": {
                "tags": [
                    "retention"
                ],
                "summary": "Generar el pdf de un Comprobante de Retencion",
                "description": "Retorna el pdf",
                "operationId": "getRetentionpdf",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Retention"
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "400": {
                        "description": "Error de Validaci\u00f3n",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValidationResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "https://facturacion.apisperu.com/api/v1"
        }
    ],
    "components": {
        "requestBodies": {
...
            "Despatch": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Despatch"
                        },
                        "examples": {
                            "example1": {
                                "$ref": "#/components/examples/remision"
                            },
                            "example2": {
                                "$ref": "#/components/examples/remisionTransportista"
                            }
                        }
                    }
                },
                "description": "Comprobante",
                "required": true
            },
            "Retention": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Retention"
                        },
                        "examples": {
                            "example1": {
                                "$ref": "#/components/examples/retencion"
                            }
                        }
                    }
                },
                "description": "Comprobante",
                "required": true
            },
...
        },
        "examples": {
...
            "remision": {
                "summary": "Gu\u00eda de Remisi\u00f3n",
                "value": {
                    "version": 2022,
                    "tipoDoc": "09",
                    "serie": "T001",
                    "correlativo": "123",
                    "fechaEmision": "2024-01-17T00:00:00-05:00",
                    "company": {
                        "ruc": 20000000001,
                        "razonSocial": "Mi empresa",
                        "nombreComercial": "Mi empresa",
                        "address": {
                            "direccion": "Direccion empresa",
                            "provincia": "LIMA",
                            "departamento": "LIMA",
                            "distrito": "LIMA",
                            "ubigueo": "150101"
                        }
                    },
                    "destinatario": {
                        "tipoDoc": "6",
                        "numDoc": 20000000002,
                        "rznSocial": "EMPRESA A"
                    },
                    "tercero": {
                        "tipoDoc": "6",
                        "numDoc": 20000000003,
                        "rznSocial": "EMPRESA SA"
                    },
                    "observacion": "NOTA GUIA",
                    "docBaja": {
                        "tipoDoc": "09",
                        "nroDoc": "T001-00001"
                    },
                    "relDoc": {
                        "tipoDoc": "02",
                        "nroDoc": "213123"
                    },
                    "envio": {
                        "codTraslado": "01",
                        "desTraslado": "VENTA",
                        "modTraslado": "01",
                        "fecTraslado": "2024-01-17T00:00:00-05:00",
                        "fecEntregaBienes": "2024-01-17T00:00:00-05:00",
                        "codPuerto": "123",
                        "indTransbordo": false,
                        "pesoTotal": 12.5,
                        "undPesoTotal": "KGM",
                        "numContenedor": "XD-2232",
                        "llegada": {
                            "ubigueo": "150101",
                            "direccion": "AV LIMA"
                        },
                        "partida": {
                            "ubigueo": "150203",
                            "direccion": "AV ITALIA"
                        },
                        "transportista": {
                            "tipoDoc": "6",
                            "numDoc": "20000000002",
                            "rznSocial": "TRANSPORTES S.A.C",
                            "placa": "ABI-453",
                            "choferTipoDoc": "1",
                            "choferDoc": "40003344"
                        }
                    },
                    "details": [
                        {
                            "cantidad": 2,
                            "unidad": "ZZ",
                            "descripcion": "PRODUCTO 1",
                            "codigo": "PROD1"
                        }
                    ]
                }
            },
            "remisionTransportista": {
                "summary": "Gu\u00eda de Remisi\u00f3n Transportista",
                "value": {
                    "version": 2022,
                    "tipoDoc": "31",
                    "serie": "V001",
                    "correlativo": "123",
                    "fechaEmision": "2026-06-23T00:00:00-05:00",
                    "company": {
                        "ruc": 20000000001,
                        "razonSocial": "Mi empresa",
                        "nombreComercial": "Mi empresa",
                        "address": {
                            "direccion": "Direccion empresa",
                            "provincia": "LIMA",
                            "departamento": "LIMA",
                            "distrito": "LIMA",
                            "ubigueo": "150101"
                        }
                    },
                    "destinatario": {
                        "tipoDoc": "6",
                        "numDoc": 20000000002,
                        "rznSocial": "EMPRESA A"
                    },
                    "observacion": "detalles",
                    "envio": {
                        "codTraslado": "01",
                        "desTraslado": "VENTA",
                        "modTraslado": "01",
                        "fecTraslado": "2026-06-24T00:00:00-05:00",
                        "fecEntregaBienes": "2026-06-24T00:00:00-05:00",
                        "codPuerto": "123",
                        "indTransbordo": false,
                        "pesoTotal": 12.5,
                        "undPesoTotal": "KGM",
                        "numContenedor": "XD-2232",
                        "llegada": {
                            "ubigueo": "150101",
                            "direccion": "AV LIMA"
                        },
                        "partida": {
                            "ubigueo": "150203",
                            "direccion": "AV ITALIA"
                        },
                        "transportista": {
                            "tipoDoc": "6",
                            "numDoc": "20000000002",
                            "rznSocial": "TRANSPORTES S.A.C",
                            "placa": "ABI-453",
                            "choferTipoDoc": "1",
                            "choferDoc": "40003344"
                        },
                        "vehiculo": {
                            "placa": "ABITR564"
                        },
                        "choferes": [
                            {
                                "tipo": "Principal",
                                "tipoDoc": "1",
                                "nroDoc": "79214014",
                                "nombres": "lUIS",
                                "apellidos": "Garcia",
                                "licencia": "M46545654"
                            }
                        ]
                    },
                    "remitente": {
                        "tipoDoc": "6",
                        "numDoc": "20000000003",
                        "rznSocial": "EMPRESA B"
                    },
                    "details": [
                        {
                            "cantidad": 2,
                            "unidad": "ZZ",
                            "descripcion": "PRODUCTO 1",
                            "codigo": "PROD1"
                        }
                    ]
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Value: Bearer {jwt}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "Retention": {
                "type": "object",
                "properties": {
                    "serie": {
                        "type": "string"
                    },
                    "correlativo": {
                        "type": "string"
                    },
                    "fechaEmision": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "proveedor": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "company": {
                        "$ref": "#/components/schemas/Company"
                    },
                    "regimen": {
                        "type": "string"
                    },
                    "tasa": {
                        "type": "number",
                        "format": "float"
                    },
                    "impRetenido": {
                        "type": "number",
                        "format": "float"
                    },
                    "impPagado": {
                        "type": "number",
                        "format": "float"
                    },
                    "observacion": {
                        "type": "string"
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RetentionDetail"
                        }
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
...
            "Despatch": {
                "type": "object",
                "properties": {
                    "version": {
                        "type": "string"
                    },
                    "tipoDoc": {
                        "description": "<a href='http://www.sunat.gob.pe/legislacion/superin/2017/anexoVII-117-2017.pdf' target='_blank'>Cat\u00e1logo No. 01</a>",
                        "type": "string"
                    },
                    "serie": {
                        "type": "string"
                    },
                    "correlativo": {
                        "type": "string"
                    },
                    "observacion": {
                        "type": "string"
                    },
                    "fechaEmision": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "company": {
                        "$ref": "#/components/schemas/Company"
                    },
                    "destinatario": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "tercero": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "comprador": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "envio": {
                        "$ref": "#/components/schemas/Shipment"
                    },
                    "remitente": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "docBaja": {
                        "$ref": "#/components/schemas/Document"
                    },
                    "relDoc": {
                        "$ref": "#/components/schemas/Document"
                    },
                    "addDocs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AdditionalDoc"
                        }
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DespatchDetail"
                        }
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "CdrResponse": {
                "type": "object",
                "properties": {
                    "accepted": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "id": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "SummaryResult": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "error": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "ticket": {
                        "type": "string"
                    }
                }
            },
            "StatusResult": {
                "type": "object",
                "properties": {
                    "cdrZip": {
                        "type": "string",
                        "format": "byte"
                    },
                    "cdrResponse": {
                        "$ref": "#/components/schemas/CdrResponse"
                    },
                    "success": {
                        "type": "boolean"
                    },
                    "error": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "code": {
                        "type": "string"
                    }
                }
            },
            "DocumentResponse": {
                "type": "object",
                "properties": {
                    "xml": {
                        "type": "string"
                    },
                    "hash": {
                        "type": "string",
                        "description": "Resumen de la Firma Digital"
                    },
                    "sunatResponse": {
                        "$ref": "#/components/schemas/BillResult"
                    }
                }
            },
            "SummaryResponse": {
                "type": "object",
                "properties": {
                    "xml": {
                        "type": "string"
                    },
                    "hash": {
                        "type": "string",
                        "description": "Resumen de la Firma Digital"
                    },
                    "sunatResponse": {
                        "$ref": "#/components/schemas/SummaryResult"
                    }
                }
            },
            "ValidationResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "field": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}